[JSC] Fix `ToIndex(value)` to Align with TC39 spec https://bugs.webkit.org/show_bug.cgi?id=298894
Reviewed by Darin Adler and Yusuke Suzuki.

This patch fixes `ToIndex(value)` implementation to align with TC39 [1].
This fix indirectly corrects functions that rely on `ToIndex(value)`,
such as
- `BigInt.asUintN` [2]
- `BigInt.asIntN` [3]
- `GetViewValue` [4]
- `SetViewValue` [5]
- `ValidateAtomicAccess` [6]

[1]: https://tc39.es/ecma262/#sec-toindex
[2]: https://tc39.es/ecma262/#sec-bigint.asuintn
[3]: https://tc39.es/ecma262/#sec-bigint.asintn
[4]: https://tc39.es/ecma262/#sec-getviewvalue
[5]: https://tc39.es/ecma262/#sec-setviewvalue
[6]: https://tc39.es/ecma262/#sec-validateatomicaccess

* JSTests/stress/SharedArrayBuffer.js:
* JSTests/stress/bigint-asintn.js:
(shouldThrow):
* JSTests/stress/bigint-asuintn.js:
(shouldThrow):
* JSTests/stress/data-view-bigint.js:
(shouldBe):
(shouldThrow):
* Source/JavaScriptCore/runtime/AtomicsObject.cpp:
(JSC::atomicsWaitImpl):
(JSC::JSC_DEFINE_HOST_FUNCTION):
(JSC::getWaiterListSize):
* Source/JavaScriptCore/runtime/BigIntConstructor.cpp:
(JSC::JSC_DEFINE_HOST_FUNCTION):
* Source/JavaScriptCore/runtime/JSCJSValue.h:
* Source/JavaScriptCore/runtime/JSCJSValueInlines.h:
(JSC::JSValue::toIndex const):
* Source/JavaScriptCore/runtime/JSDataViewPrototype.cpp:
(JSC::getData):
(JSC::setData):
* Source/JavaScriptCore/runtime/JSGenericTypedArrayView.h:
* Source/JavaScriptCore/runtime/JSGenericTypedArrayViewInlines.h:
(JSC::JSGenericTypedArrayView<Adaptor>::inBounds const):

Canonical link: https://commits.webkit.org/300157@main

+++ Source/JavaScriptCore/runtime/JSGenericTypedArrayView.h
@@ -109,7 +109,7 @@ public:
     std::span<const typename Adaptor::Type> typedSpan() const { return unsafeMakeSpan(typedVector(), length()); }
     std::span<typename Adaptor::Type> typedSpan() { return unsafeMakeSpan(typedVector(), length()); }
 
-    inline bool inBounds(size_t) const;
+    inline bool inBounds(uint64_t) const;
 
     // These methods are meant to match indexed access methods that JSObject
     // supports - hence the slight redundancy.
